pip install -r requirements.txt
pandas
plotly
ollama
requests
pygithub
streamlit
